pyproject.toml
[project]
name = "omni-retrieval"
version = "0.1.0"
description = "Modular RAG pipeline with swappable stages"
requires-python = ">=3.10"
readme = "README.md"
dependencies = [
	"llama-index-core",
	"llama-index-llms-openai-like",
	"llama-index-embeddings-openai",
	"llama-index-vector-stores-qdrant",
	"llama-index-readers-file",
	"qdrant-client",
	"pyyaml",
	"pydantic",
	"python-dotenv",
	"click",
	"rich",
	"datasets",
	"fastembed",
	"dashscope",
	"ragas",
	"llama-index-retrievers-bm25",
]

[dependency-groups]
dev = [
	"ruff",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
